[firebase][config]-Setup config for firestore and fcm
