+++ crypto/src/main/java/itu/crypto/firebase/config/FirebaseConfiguration.java
@@ -1,10 +1,8 @@
-package itu.crypto.firebase.firestore;
+package itu.crypto.firebase.config;
 
 import com.google.auth.oauth2.GoogleCredentials;
-import com.google.cloud.firestore.Firestore;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
-import com.google.firebase.cloud.FirestoreClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -15,20 +13,19 @@ import java.io.IOException;
 import java.io.InputStream;
 
 @Configuration
-public class FirestoreConfiguration {
+public class FirebaseConfiguration {
 
     @Value("classpath:/private-key.json")
     private Resource privateKey;
 
     @Bean
-    public Firestore firestore() throws IOException {
+    public FirebaseApp firebaseApp() throws IOException {
         InputStream credentials = new ByteArrayInputStream(privateKey.getContentAsByteArray());
 
         FirebaseOptions firebaseOptions = FirebaseOptions.builder()
                 .setCredentials(GoogleCredentials.fromStream(credentials))
                 .build();
 
-        FirebaseApp firebaseApp = FirebaseApp.initializeApp(firebaseOptions);
-        return FirestoreClient.getFirestore(firebaseApp);
+        return FirebaseApp.initializeApp(firebaseOptions);
     }
 }
