+++ crypto/src/main/java/itu/crypto/firebase/config/FirebaseMessagingConfiguration.java
@@ -1,0 +1,21 @@
+package itu.crypto.firebase.config;
+
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.messaging.FirebaseMessaging;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.io.IOException;
+
+@Configuration
+@RequiredArgsConstructor
+public class FirebaseMessagingConfiguration {
+
+    private final FirebaseApp firebaseApp;
+
+    @Bean
+    public FirebaseMessaging firebaseMessaging() throws IOException {
+        return FirebaseMessaging.getInstance(firebaseApp);
+    }
+}
