+++ crypto/src/main/java/itu/crypto/firebase/config/FirestoreConfiguration.java
@@ -1,0 +1,22 @@
+package itu.crypto.firebase.config;
+
+import com.google.cloud.firestore.Firestore;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.cloud.FirestoreClient;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.io.IOException;
+
+@Configuration
+@RequiredArgsConstructor
+public class FirestoreConfiguration {
+
+    private final FirebaseApp firebaseApp;
+
+    @Bean
+    public Firestore firestore() throws IOException {
+        return FirestoreClient.getFirestore(firebaseApp);
+    }
+}
